feat: add search by name with queryParams
